Add image to user results sheet

+++ centr_osvita/quiz/signals.py
@@ -32,5 +32,5 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
 
     new_image = instance.image
     if not old_image == new_image:
-        if os.path.isfile(old_image.path):
+        if old_image and os.path.isfile(old_image.path):
             os.remove(old_image.path)
